support for float expenses

+++ app/src/main/java/ajitsingh/com/expensemanager/activity/CurrentMonthExpenseFragment.java
@@ -44,7 +44,7 @@ public class CurrentMonthExpenseFragment extends Fragment implements CurrentMont
   }
 
   @Override
-  public void displayTotalExpense(Long totalExpense) {
+  public void displayTotalExpense(Float totalExpense) {
     TextView totalExpenseTextBox = (TextView) getActivity().findViewById(R.id.current_months_total_expense);
     totalExpenseTextBox.setText(getString(R.string.total_expense) + " " + getString(R.string.rupee_sym) + totalExpense);
   }
