+++ app/src/main/java/ajitsingh/com/expensemanager/activity/TodaysExpenseFragment.java
@@ -36,7 +36,7 @@ public class TodaysExpenseFragment extends Fragment implements TodaysExpenseView
   }
 
   @Override
-  public void displayTotalExpense(Long totalExpense) {
+  public void displayTotalExpense(Float totalExpense) {
     TextView totalExpenseTextBox = (TextView) getActivity().findViewById(R.id.total_expense);
     totalExpenseTextBox.setText(getActivity().getString(R.string.total_expense) + " " + getActivity().getString(R.string.rupee_sym) + totalExpense.toString());
   }
