+++ app/src/main/java/ajitsingh/com/expensemanager/view/CurrentMonthExpenseView.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface CurrentMonthExpenseView {
   void displayGraph(List<Bar> points);
 
-  void displayTotalExpense(Long totalExpense);
+  void displayTotalExpense(Float totalExpense);
 
   int getGraphColor();
 }
